+++ rest/describers/doc.go
@@ -1,0 +1,3 @@
+/* Package describers holds different sources of restful scenarios
+ */
+package describers