+++ rest/doc.go
@@ -1,0 +1,7 @@
+/*Package rest adds capabilities for testing restful services.
+
+It should not have any assertions itself--those belong in the should
+package. Rather can provide of series of requests and responses ready
+for testing and comparison.
+*/
+package rest